requirements.txt
pandas
numpy
streamlit
Pillow
pydeck